requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==41.0.7
Django==5.0
djangorestframework==3.14.0
jwt==1.3.1
Pillow==10.1.0
psycopg2==2.9.9
pycparser==2.21
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.4